+++ src/main/java/ladder/domain/GameConsole.java
@@ -1,5 +1,12 @@
 package ladder.domain;
 
+import ladder.domain.ladder.Ladder;
+import ladder.domain.participants.Participants;
+import ladder.domain.participants.Person;
+import ladder.domain.result.GameResult;
+import ladder.domain.result.Result;
+import ladder.domain.result.Results;
+
 import java.util.LinkedHashMap;
 import java.util.Map;
