+++ src/main/java/ladder/domain/ladder/Ladder.java
@@ -1,8 +1,7 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Random;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
