+++ src/main/java/ladder/domain/ladder/Point.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 public class Point {
     private static final int NEXT = 1;
