+++ src/main/java/ladder/domain/ladder/PointGenerator.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 @FunctionalInterface
 public interface PointGenerator {
