+++ src/main/java/ladder/domain/participants/Participants.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.participants;
 
 import java.util.Arrays;
 import java.util.Collections;
