+++ src/main/java/ladder/domain/participants/Person.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.participants;
 
 import java.util.Objects;
