+++ src/test/java/ladder/domain/GameConsoleTest.java
@@ -1,5 +1,8 @@
 package ladder.domain;
 
+import ladder.domain.ladder.Ladder;
+import ladder.domain.participants.Participants;
+import ladder.domain.result.Results;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
