+++ src/test/java/ladder/domain/participants/PersonTest.java
@@ -1,5 +1,6 @@
-package ladder.domain;
+package ladder.domain.participants;
 
+import ladder.domain.participants.Person;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
