+++ src/test/java/ladder/domain/result/GameResultTest.java
@@ -1,5 +1,8 @@
-package ladder.domain;
+package ladder.domain.result;
 
+import ladder.domain.participants.Person;
+import ladder.domain.result.GameResult;
+import ladder.domain.result.Result;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
