+++ src/test/java/ladder/domain/result/ResultsTest.java
@@ -1,5 +1,6 @@
-package ladder.domain;
+package ladder.domain.result;
 
+import ladder.domain.result.Results;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
